Adds a cast to print file sizes (intmax_t), to make happy both Linux and Darwin.
git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@2598 a4d8336d-3463-0410-8bba-c098c45d37a8

+++ resource_monitor/src/resource_monitor.c
@@ -937,7 +937,7 @@ void monitor_wds_summary_once(struct wdir_info *d, int divisor)
 
 	// files+directories byte_count
 	fprintf(log_file, "%10d\t", (d->files + d->directories) / divisor);
-	fprintf(log_file, "%10jd\t", d->byte_count / divisor);
+	fprintf(log_file, "%10jd\t", (intmax_t) d->byte_count / divisor);
 }
 
 void monitor_fss_summary_once(struct filesys_info *f, int divisor)
